Guice: Use child injector instead of scoping. Fixes #132.
