+++ src/main/java/org/granitepowered/granite/impl/guice/ConfigDirAnnotation.java
@@ -23,16 +23,45 @@
 
 package org.granitepowered.granite.impl.guice;
 
-import com.google.inject.ScopeAnnotation;
+import org.spongepowered.api.service.config.ConfigDir;
 
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
+import java.lang.annotation.Annotation;
 
-@Target({ElementType.TYPE, ElementType.METHOD})
-@Retention(RetentionPolicy.RUNTIME)
-@ScopeAnnotation
-public @interface PluginScoped {
+class ConfigDirAnnotation implements ConfigDir {
+
+    private final boolean isShared;
+
+    public ConfigDirAnnotation(boolean isShared) {
+        this.isShared = isShared;
+    }
+
+    @Override
+    public boolean sharedRoot() {
+        return isShared;
+    }
+
+    @Override
+    public Class<? extends Annotation> annotationType() {
+        return ConfigDir.class;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) {
+            return true;
+        }
+
+        if (o == null || !(o instanceof ConfigDir)) {
+            return false;
+        }
+
+        ConfigDir that = (ConfigDir) o;
+        return sharedRoot() == that.sharedRoot();
+    }
+
+    @Override
+    public int hashCode() {
+        return (127 * "sharedRoot".hashCode()) ^ Boolean.valueOf(sharedRoot()).hashCode();
+    }
 
 }
