--- src/main/java/org/granitepowered/granite/impl/guice/PluginScope.java
@@ -1,96 +1,0 @@
-/*
- * License (MIT)
- *
- * Copyright (c) 2014-2015 Granite Team
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of this
- * software and associated documentation files (the "Software"), to deal in the
- * Software without restriction, including without limitation the rights to use, copy,
- * modify, merge, publish, distribute, sublicense, and/or sell copies of the Software,
- * and to permit persons to whom the Software is furnished to do so, subject to the
- * following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
- * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
- * PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
- * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
- * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
- * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
-
-package org.granitepowered.granite.impl.guice;
-
-import static com.google.common.base.Preconditions.checkState;
-
-import com.google.common.collect.Maps;
-import com.google.inject.Key;
-import com.google.inject.OutOfScopeException;
-import com.google.inject.Provider;
-import com.google.inject.Scope;
-import org.spongepowered.api.plugin.PluginContainer;
-
-import java.util.Map;
-
-public class PluginScope implements Scope {
-
-    private final ThreadLocal<Map<Key<?>, Object>> values = new ThreadLocal<>();
-
-    public void enter(PluginContainer container) {
-        checkState(values.get() == null, "A scoping block is already in progress");
-        values.set(Maps.<Key<?>, Object>newHashMap());
-        seed(PluginContainer.class, container);
-    }
-
-    public void exit() {
-        checkState(values.get() != null, "No scoping block in progress");
-        values.remove();
-    }
-
-    public <T> void seed(Key<T> key, T value) {
-        Map<Key<?>, Object> scopedObjects = getScopedObjectMap(key);
-        checkState(!scopedObjects.containsKey(key), "A value for the key %s was already seeded in this scope. Old value: %s New value: %s", key,
-                   scopedObjects.get(key), value);
-        scopedObjects.put(key, value);
-    }
-
-    public <T> void seed(Class<T> clazz, T value) {
-        seed(Key.get(clazz), value);
-    }
-
-    @SuppressWarnings("unchecked")
-    public <T> T getInstance(Key<T> key) {
-        Map<Key<?>, Object> scopedObjects = getScopedObjectMap(key);
-        return (T) scopedObjects.get(key);
-    }
-
-    public <T> Provider<T> scope(final Key<T> key, final Provider<T> unscoped) {
-        return new Provider<T>() {
-            public T get() {
-                Map<Key<?>, Object> scopedObjects = getScopedObjectMap(key);
-
-                @SuppressWarnings("unchecked")
-                T current = (T) scopedObjects.get(key);
-
-                if (current == null && !scopedObjects.containsKey(key)) {
-                    current = unscoped.get();
-
-                    scopedObjects.put(key, current);
-                }
-
-                return current;
-            }
-        };
-    }
-
-    private <T> Map<Key<?>, Object> getScopedObjectMap(Key<T> key) {
-        Map<Key<?>, Object> scopedObjects = values.get();
-        if (scopedObjects == null) {
-            throw new OutOfScopeException("Cannot access " + key + " outside of a scoping block");
-        }
-        return scopedObjects;
-    }
-
-}
